Split hook
